+++ apps/headless-crawler/functions/E-JobDetailHandler/handler.ts
@@ -1,7 +1,7 @@
 import type { SQSEvent, SQSHandler } from "aws-lambda";
 import { Effect, Exit } from "effect";
 import { extractJobDetailRawHtml } from "../../lib/jobDetailExtractor";
-import { fromEventToFirstRecord } from "../extractTransformAndSaveJobDetailHandler/helper";
+import { fromEventToFirstRecord } from "../E-T-L-JobDetailHandler/helper";
 
 export const handler: SQSHandler = async (event: SQSEvent) => {
   const effect = Effect.gen(function* () {
