+++ apps/headless-crawler/functions/E-T-L-JobDetailHandler/handler.ts
