+++ apps/headless-crawler/functions/E-T-L-JobDetailHandler/helper.ts
