+++ apps/headless-crawler/functions/E-T-L-JobDetailHandler/schema.ts
