+++ apps/headless-crawler/functions/E-T-L-JobDetailHandler/type.ts
