+++ apps/headless-crawler/infra/constructs/E-jobDetail.ts
@@ -17,7 +17,7 @@ export class JobDetailRawHtmlExtractorConstruct extends Construct {
   ) {
     super(scope, id);
     this.extractor = new NodejsFunction(this, id, {
-      entry: "functions/extractJobDetailHandler/handler.ts",
+      entry: "functions/E-JobDetailHandler/handler.ts",
       handler: "handler",
       runtime: lambda.Runtime.NODEJS_22_X,
       memorySize: 1024,
